refactoring : created useMode hook. Replaced all redux to top parent component ModeSwitcher.jsx
